+++ app/src/main/res/values/strings.xml
@@ -7,18 +7,37 @@
     <string name="temporal_btn">Botón</string>
     <string name="temporal_proceso">Haciendo Algo...</string>
 
+    <!-- Crear y modificar gastos -->
     <string name="titulo_nuevo_gasto">Agregar un gasto</string>
     <string name="titulo_editar_gasto">Modificar un gasto</string>
 
+    <!-- Comunes -->
+    <string name="confirmar">Confirmar</string>
+    <string name="cancelar">Cancelar</string>
+
     <!-- Notificaciones -->
+    <string name="id_canal_notif">canal_recordatorios</string>
     <string name="nombre_canal_notif">Recordatorios de presupuesto</string>
-    <string name="descripcion_canal_notif">Envvía una notificación cuando excedes tu presupuesto diario.</string>
+    <string name="descripcion_canal_notif">
+        Envvía una notificación cuando excedes tu presupuesto diario.
+    </string>
 
     <string name="titulo_notificacion_presupuesto">Alerta de Presupuesto</string>
     <string name="contenido_notificacion_presupuesto">Superaste tu presupuesto diario por $</string>
+
+
+    <!-- Activity Presupuesto Diario -->
+    <string name="titulo_definir_presupuesto">Presupuesto diario</string>
+    <string name="descripcion_definir_presupuesto">¿Cuánto quieres gastar por día?</string>
+    <string name="hint_definir_presupuesto">$200.00</string>
+
     <string-array name="tipos_de_gasto">
         <item>Necesario</item>
         <item>Entretenimiento</item>
         <item>Extra</item>
     </string-array>
+
+    <!-- Keys de SharedPreferences -->
+    <string name="pref_key_email_usuario">EMAIL_USUARIO</string>
+    <string name="pref_key_password_usuario">PASSWORD_USUARIO</string>
 </resources>
